Create Chapter_2.md
I hope I did it right, but I wanted to try and help out. 
Please let me know if this is the right way to format a chapter. I haven't changed any words yet but I wanted to be sure it works and all first. 
I'm brand new to github and thereby the coding world, but I do like writing tho... 
So yeah... hope this is okay and if not please let me know how to do it the right way